ready for training discrete env
