--- haptic/video/openaigym.episode_batch.0.98750.stats.json
@@ -1,1 +1,0 @@
-{"initial_reset_timestamp": 10780.960714067, "timestamps": [10870.80247832, 10908.190775516, 10932.810409305, 10978.444025383, 11022.573534744, 11033.867473756, 11076.323141246, 11121.095565017, 11231.877842367, 11277.668737146, 11324.206858195], "episode_lengths": [5000, 1840, 2967, 5000, 5000, 1116, 5000, 5000, 5000, 5000, 5000], "episode_rewards": [-435.37414965989353, -219.4932203389808, -321.866903914583, -387.3239436619967, -362.0689655172566, 73.05284552845629, -426.5734265734557, -427.15231788083065, -424.8120300752147, -473.5099337748739, -393.93939393941235], "episode_types": ["t", "t", "t", "t", "t", "t", "t", "t", "t", "t", "t", "t"]}
