--- haptic/video/openaigym.manifest.0.98750.manifest.json
@@ -1,1 +1,0 @@
-{"stats": "openaigym.episode_batch.0.98750.stats.json", "videos": [["openaigym.video.0.98750.video000000.mp4", "openaigym.video.0.98750.video000000.meta.json"], ["openaigym.video.0.98750.video000001.mp4", "openaigym.video.0.98750.video000001.meta.json"], ["openaigym.video.0.98750.video000008.mp4", "openaigym.video.0.98750.video000008.meta.json"]], "env_info": {"gym_version": "0.21.0"}}
