+++ tests/test_main.py
@@ -1,12 +1,12 @@
-from haptic.__main__ import add
+# from haptic.__main__ import add
 
 # from haptic.env.car_racer_env import subtract
 
 # from haptic.env.car_racer_env import HapticCarRacer
 
 
-def test_add():
-    assert add(2, 3) == 5
+# def test_add():
+#     assert add(2, 3) == 5
 
 
 # def test_subtract():
